+++ HttPeek/Classes/WebViewPeek.mm
@@ -14,7 +14,7 @@ WebViewDelegate *_webViewDelegate;
 - (BOOL)webView:(UIWebView *)webView shouldStartLoadWithRequest:(NSURLRequest *)request navigationType:(UIWebViewNavigationType)navigationType;
 {
 	_LogRequest(request);
-	NSLog(@"%@ shouldStartLoadWithRequest: %@, navigationType:%d", webView, [request.URL.absoluteString stringByReplacingPercentEscapesUsingEncoding:NSUTF8StringEncoding], navigationType);
+	NSLog(@"%@ shouldStartLoadWithRequest: %@, navigationType:%d", webView, [request.URL.absoluteString stringByReplacingPercentEscapesUsingEncoding:NSUTF8StringEncoding], (int)navigationType);
 	id<UIWebViewDelegate> delegate = [_delegates objectForKey:[NSString stringWithFormat:@"%p", webView]];
 	return [delegate respondsToSelector:@selector(webView: shouldStartLoadWithRequest: navigationType:)] ? [delegate webView:webView shouldStartLoadWithRequest:request navigationType:navigationType] : YES;
 }
